Address feedback on Meshopt implementation
- Add CC BY Attribution notice
- Consolidate files
- Styleguide fixes
