--- packages/model-viewer/src/types/meshoptimizer.ts
@@ -1,10 +1,0 @@
-interface MeshoptDecoder {
-    ready: Promise<void>;
-    supported: boolean;
-}
-
-declare global {
-    const MeshoptDecoder: MeshoptDecoder;
-}
-
-export {};
