--- packages/model-viewer/src/utilities/fetch-script.ts
@@ -1,10 +1,0 @@
-export function fetchScript(src: string): Promise<Event> {
-    return new Promise((resolve, reject) => {
-        const script = document.createElement('script');
-        document.body.appendChild(script);
-        script.onload = resolve;
-        script.onerror = reject;
-        script.async = true;
-        script.src = src;
-    });
-}
